improve error handling

+++ src/app/pages/Logout.jsx
@@ -2,33 +2,47 @@ import React from 'react';
 import { useHistory, Redirect } from 'react-router-dom';
 import { Heading, Spinner, Paragraph, Pane, majorScale } from 'evergreen-ui';
 
-import GenericActionFullscreen from '../../library/components/GenericActionFullscreen';
 import AuthenticationService from '../../library/services/AuthenticationService';
+import { useOperation } from '../../library/services/ReactGenericOperation';
+
 import Semantics from '../utils/semantics';
-import FullScreenError from '../components/FullscreenError';
+import FullscreenCenterContainer from '../components/FullscreenCenterContainer';
+import FullscreenError from '../components/FullscreenError';
+import { useAppContextState } from '../context/AppContext'
 
 function Logout() {
   if (!AuthenticationService.isAuthenticated()) {
     return <Redirect to="/" />
   }
   
-  return <LogoutDialog />
+  return <FullscreenCenterContainer><LogoutDialog /></FullscreenCenterContainer>
 };
 
 const LogoutDialog = () => {
   const history = useHistory();
+  const { genericOperationBuilder } = useAppContextState()
 
-  return <GenericActionFullscreen
-    actionKey={Semantics.vnd_jeera.actions.logout}
-    Loading={<Loading />}
-    Success={Success}
-    ErrorComponent={FullScreenError}
-    Component={Component}
-    onSuccessCallback={() => {
-      AuthenticationService.removeToken();
-      setTimeout(() => history.push('/'), 2000);
-    }}
-  />
+  const logoutOperation = genericOperationBuilder.fromKey(Semantics.vnd_jeera.actions.logout)
+  const { parametersDetail, makeCall, isLoading, success, data, error } = useOperation(logoutOperation)
+
+  if (isLoading) {
+    return <Loading />
+  } else if (error) {
+    return <FullscreenError error={error}/>
+  } else if (success) {
+    AuthenticationService.removeToken();
+    setTimeout(() => history.push('/'), 2000);
+    return <Success data={data}/>
+  } else if (parametersDetail.documentation.length === 0) {
+    // ({error, triggerCall, filtersToDisplay, formToDisplay})
+    if (makeCall !== undefined) makeCall()
+    return (<>
+      <Heading width="100%" size={700} marginBottom={majorScale(2)}>We are logging you out...</Heading>
+      { error && <Paragraph width="100%" size={500}>{error}</Paragraph> }
+    </>)
+  } else {
+    return <Heading>Not yet supported</Heading>
+  }
 }
 
 const Loading = () =>
@@ -43,17 +57,5 @@ const Success = ({ data }) =>
     <Paragraph>We will redirect you to the home page very soon.</Paragraph>
     { data && <Paragraph>{JSON.stringify(data.data)}</Paragraph> }
   </>
-
-const Component = ({error, triggerCall, filtersToDisplay, formToDisplay}) => {
-  if (!filtersToDisplay && !formToDisplay) {
-    if (triggerCall !== undefined) triggerCall();
-    return (<>
-      <Heading width="100%" size={700} marginBottom={majorScale(2)}>We are logging you out...</Heading>
-      { error && <Paragraph width="100%" size={500}>{error}</Paragraph> }
-    </>)
-  } else {
-    return <Heading>Not yet supported</Heading>
-  }
-}
 
 export default Logout;
